feat: Add GitHub Actions workflow for building and deploying
This commit adds a new GitHub Actions workflow file, deploy_gh_pages.yaml, which is responsible for building and deploying the project. The workflow is triggered on pushes to the master and main branches. It uses Ubuntu as the runner and performs the following steps:
- Checkout the repository
- Install Python 3.8
- Install Julia 1.5
- Install necessary dependencies for pre-rendering
- Build and deploy the site using the JamesIves/github-pages-deploy-action

The workflow will deploy the site to the gh-pages branch and the generated site files will be stored in the __site folder.